ITS WORKING :D Trig på manuel værdi af trans og offset og std på PDH

+++ Arduino/libraries/PID_Bach/PID_Bach_v1.h
@@ -71,7 +71,7 @@ class PID
     uint16_t *myVerboseNum;
     int *DirecPoin;
     bool *AutoPoin;
-    int *HillPoin;
+    float *HillPoin;
 
     int outputSum;
     int lastInput;
